requirements.txt
pyyaml
numpy
pandas
torch
scikit-learn

transformers>=4.57.0
datasets>=2.18.0
accelerate>=0.28.0
